WIP: strip weekly-workflow tree; keep the smoke + proxy
Cuts everything that only existed to support the env-gated-off weekly
Temporal schedule. The scope of this PR is now exactly: OAuth-gated
ClickHouse proxy + vendored pi-autoresearch + smoke command that drives
a local Docker sandbox. No Task creation, no fan-out workflow, no Slack
summary, no PR-writing handoff.

Deleted:
- products/query_performance_ai/backend/management/commands/register_weekly_autoresearch_schedule.py
- products/query_performance_ai/backend/temporal/{schedule,workflows,activities,__init__}.py
- products/query_performance_ai/backend/slow_queries.py
- products/query_performance_ai/backend/prompts/{code_handoff.py,__init__.py,templates/code_handoff.md}
- products/query_performance_ai/backend/tests/test_{code_handoff,slow_queries_and_activities}.py
- products/tasks/backend/temporal/process_task/activities/run_autoresearch_campaign.py
- products/tasks/backend/migrations/0031_alter_task_origin_product.py
- .agents/skills/verifying-clickhouse-perf-fixes/ (no PR-writing phase to gate)

Moved _harvest_artifacts + its tuple/list helpers +
RunAutoresearchCampaignOutput dataclass + LLM_GATEWAY_PRODUCT_SLUG from
the deleted tasks-activity into a clean helper module at
products/query_performance_ai/backend/harvest.py so the smoke doesn't
depend on Temporal / Task machinery.

Edited:
- products/tasks/backend/models.py — drop OriginProduct.QUERY_PERFORMANCE
- products/tasks/backend/migrations/max_migration.txt → 0030_taskautomation
- products/tasks/backend/temporal/{__init__,process_task/activities/__init__,process_task/workflow}.py — unwire
- posthog/management/commands/start_temporal_worker.py — no QUERY_PERF_AI_* imports/registration
- run_autoresearch_smoke.py — import from the new harvest module

Tests: 14 passing (proxy + transports only, everything else went with
the weekly tree).

Generated-By: PostHog Code
Task-Id: 8e68715d-6e5c-4085-88f3-a0809bf69dac
