+++ posthog/management/commands/start_temporal_worker.py
@@ -168,10 +168,6 @@ from products.signals.backend.temporal import (
     ACTIVITIES as SIGNALS_PRODUCT_ACTIVITIES,
     WORKFLOWS as SIGNALS_PRODUCT_WORKFLOWS,
 )
-from products.query_performance_ai.backend.temporal import (
-    ACTIVITIES as QUERY_PERF_AI_ACTIVITIES,
-    WORKFLOWS as QUERY_PERF_AI_WORKFLOWS,
-)
 from products.tasks.backend.temporal import (
     ACTIVITIES as TASKS_ACTIVITIES,
     WORKFLOWS as TASKS_WORKFLOWS,
@@ -252,8 +248,8 @@ _task_queue_specs = [
     ),
     (
         settings.TASKS_TASK_QUEUE,
-        TASKS_WORKFLOWS + QUERY_PERF_AI_WORKFLOWS,
-        TASKS_ACTIVITIES + QUERY_PERF_AI_ACTIVITIES,
+        TASKS_WORKFLOWS,
+        TASKS_ACTIVITIES,
     ),
     (
         settings.MAX_AI_TASK_QUEUE,
