--- products/query_performance_ai/backend/prompts/__init__.py
@@ -1,3 +1,0 @@
-from .code_handoff import render_code_handoff_prompt
-
-__all__ = ["render_code_handoff_prompt"]
