+++ products/tasks/backend/temporal/__init__.py
@@ -24,7 +24,6 @@ from .process_task.activities import (
     prepare_sandbox_for_repository,
     read_sandbox_logs,
     relay_sandbox_events,
-    run_autoresearch_campaign_in_sandbox,
     send_followup_to_sandbox,
     start_agent_server,
     track_workflow_event,
@@ -53,7 +52,6 @@ ACTIVITIES = [
     forward_pending_user_message,
     relay_sandbox_events,
     create_resume_snapshot,
-    run_autoresearch_campaign_in_sandbox,
     send_followup_to_sandbox,
     start_agent_server,
     read_sandbox_logs,
